move json utils to localstack.utils.json
